requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
